% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions_documentation_v2.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Display a map by US state depicting location of accidents based on FARS data}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{Number of the US state.}

\item{year}{A single numeric or character string specifying a year.}
}
\value{
This function returns a map with the location of accidents based on
    the FARS data for a specific US state and year. The map contains a dot
    for each accident.
}
\description{
This function displays a map by US state. The map depicts the location of
    accidents based on FARS data, with a dot for each accident.
}
\examples{
fars_map_state(1, 2013)
fars_map_state(5, "2014")

}
