% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions_documentation_v2.R
\name{fars_read}
\alias{fars_read}
\title{Read csv file into tibble}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string with the name of the file.}
}
\value{
This function returns the data read in from the file as a tibble,
    a data frame with class tbl_df.
}
\description{
This function allows to read data from any csv file.
Specifically, however, the function is used to read data from the US National Highway
Traffic Safety Administration's Fatality Analysis Reporting System (FARS), which is a
nationwide census providing the American public yearly data regarding fatal
injuries suffered in motor vehicle traffic crashes.
}
\note{
The file needs to be in the current working directory.
}
\examples{
fars_read("accident_2013.csv.bz2")
\dontrun{
  fars_read("anyFile.csv")
}

}
