% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions_documentation_v2.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read files with FARS data for several years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A single numeric or character string specifying a year or
a vector of numerics or character strings of year(s).}
}
\value{
This function returns a list of tibbles. Each tibble containts FARS data
    for one year stored in individual files.
}
\description{
This function allows to read FARS data from files for one or more years.
}
\examples{
\dontrun{
  fars_read_years(2015)
  fars_read_years(c(2013, 2014))
  fars_read_years(c("2014", "2013"))
}
}
