% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions_documentation_v2.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Generate a summary of accidents per month and year from FARS data}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A single numeric or character string specifying a year or
a vector of numerics or character strings of year(s).}
}
\value{
This function returns a tibble with a row for each month and columns
    for each year. The entries in the cell correspond to the number of
    accidents in that month and year.
}
\description{
This function generates a summary of accidents per month and year from FARS data
by reading in the file(s) with the FARS data for one or more years.
}
\examples{
fars_summarize_years(2015)
fars_summarize_years(c(2013, 2014))
fars_summarize_years(c("2014", "2013"))

}
