% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions_documentation_v2.R
\name{make_filename}
\alias{make_filename}
\title{Create filename for FARS data for specific year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A numeric or character string specifying a year.}
}
\value{
This function returns a character string with the name of the file. The
    filename is created by inserting the year into the string
    "accident_\code{year}.csv.bz2".
}
\description{
This function creates a specific filename for a year to later read FARS data for
the particular year stored in a file with this filename.
}
\note{
The year needs to be between 2013 and 2015.
}
\examples{
make_filename(2014)
make_filename("2013")
}
